requirements.txt
langchain
langchain-community
transformers
sentence-transformers
qdrant-client
beautifulsoup4
pydantic
openai
huggingface-hub
jq
pandas
python-dotenv
langchain_huggingface
aiofiles
ragas